WIP/investment_app_WIP/daily_journal_app_WIP/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
python-dateutil>=2.8
Flask-Login>=0.6
python-docx>=0.8.11
pytest>=7.0
